Consolidate friendlyRelativeTime into date library
